pyproject.toml
[project]
name = "reelforge"
version = "0.1.0"
description = "An automated AI pipeline to generate Instagram Reels from ideas using AWS Bedrock and Polly."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-aws>=1.2.1",
    "moviepy>=2.2.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "boto3>=1.35.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]
